+++ docs/source/auto/line_profiler.autoprofile.ast_profle_transformer.rst
@@ -1,7 +1,7 @@
 line\_profiler.autoprofile.ast\_profle\_transformer module
 ==========================================================
 
-.. automodule:: line_profiler.autoprofile.ast_profle_transformer
+.. automodule:: line_profiler.autoprofile.ast_profile_transformer
    :members:
    :undoc-members:
    :show-inheritance:
